Move Threads::Mutex to a dedicated header.
Most files don't need anything in Threads other than Threads::Mutex so we can
decouple them from TBB et. al. by moving this class to its own header.
