+++ examples/step-48/step-48.cc
@@ -20,6 +20,7 @@
 
 // The necessary files from the deal.II library.
 #include <deal.II/base/logstream.h>
+#include <deal.II/base/multithread_info.h>
 #include <deal.II/base/utilities.h>
 #include <deal.II/base/function.h>
 #include <deal.II/base/conditional_ostream.h>
