+++ include/deal.II/base/flow_function.h
@@ -20,8 +20,8 @@
 #include <deal.II/base/config.h>
 
 #include <deal.II/base/function.h>
+#include <deal.II/base/mutex.h>
 #include <deal.II/base/point.h>
-#include <deal.II/base/thread_management.h>
 
 DEAL_II_NAMESPACE_OPEN
