+++ include/deal.II/base/function_cspline.h
@@ -20,8 +20,8 @@
 
 #ifdef DEAL_II_WITH_GSL
 #  include <deal.II/base/function.h>
+#  include <deal.II/base/mutex.h>
 #  include <deal.II/base/point.h>
-#  include <deal.II/base/thread_management.h>
 
 #  include <gsl/gsl_spline.h>
