+++ include/deal.II/fe/fe_nedelec.h
@@ -19,12 +19,12 @@
 #include <deal.II/base/config.h>
 
 #include <deal.II/base/geometry_info.h>
+#include <deal.II/base/mutex.h>
 #include <deal.II/base/polynomial.h>
 #include <deal.II/base/polynomials_nedelec.h>
 #include <deal.II/base/table.h>
 #include <deal.II/base/tensor.h>
 #include <deal.II/base/tensor_product_polynomials.h>
-#include <deal.II/base/thread_management.h>
 
 #include <deal.II/fe/fe.h>
 #include <deal.II/fe/fe_poly_tensor.h>
