+++ include/deal.II/fe/fe_raviart_thomas.h
@@ -18,6 +18,7 @@
 
 #include <deal.II/base/config.h>
 
+#include <deal.II/base/mutex.h>
 #include <deal.II/base/table.h>
 
 #include <deal.II/fe/fe.h>
