+++ include/deal.II/lac/block_matrix_base.h
@@ -20,9 +20,9 @@
 #include <deal.II/base/config.h>
 
 #include <deal.II/base/memory_consumption.h>
+#include <deal.II/base/mutex.h>
 #include <deal.II/base/smartpointer.h>
 #include <deal.II/base/table.h>
-#include <deal.II/base/thread_management.h>
 #include <deal.II/base/utilities.h>
 
 #include <deal.II/lac/block_indices.h>
@@ -33,6 +33,7 @@
 #include <deal.II/lac/vector_operation.h>
 
 #include <cmath>
+#include <mutex>
 
 DEAL_II_NAMESPACE_OPEN
 
@@ -1038,7 +1039,7 @@ private:
      * A mutex variable used to guard access to the member variables of this
      * structure;
      */
-    std::mutex mutex;
+    Threads::Mutex mutex;
 
     /**
      * Copy operator. This is needed because the default copy operator of this
