+++ include/deal.II/lac/la_parallel_vector.h
@@ -26,7 +26,6 @@
 #include <deal.II/base/parallel.h>
 #include <deal.II/base/partitioner.h>
 #include <deal.II/base/subscriptor.h>
-#include <deal.II/base/thread_management.h>
 
 #include <deal.II/lac/vector_operation.h>
 #include <deal.II/lac/vector_space_vector.h>
