+++ include/deal.II/lac/lapack_full_matrix.h
@@ -19,9 +19,9 @@
 
 #include <deal.II/base/config.h>
 
+#include <deal.II/base/mutex.h>
 #include <deal.II/base/smartpointer.h>
 #include <deal.II/base/table.h>
-#include <deal.II/base/thread_management.h>
 
 #include <deal.II/lac/lapack_support.h>
 #include <deal.II/lac/vector_memory.h>
@@ -972,7 +972,7 @@ private:
   /**
    * Thread mutex.
    */
-  mutable std::mutex mutex;
+  mutable Threads::Mutex mutex;
 };
