+++ include/deal.II/lac/scalapack.h
@@ -22,8 +22,8 @@
 
 #  include <deal.II/base/exceptions.h>
 #  include <deal.II/base/mpi.h>
+#  include <deal.II/base/mutex.h>
 #  include <deal.II/base/process_grid.h>
-#  include <deal.II/base/thread_management.h>
 
 #  include <deal.II/lac/full_matrix.h>
 #  include <deal.II/lac/lapack_full_matrix.h>
