+++ include/deal.II/matrix_free/mapping_info_storage.templates.h
@@ -19,8 +19,6 @@
 #include <deal.II/base/config.h>
 
 #include <deal.II/base/memory_consumption.h>
-#include <deal.II/base/multithread_info.h>
-#include <deal.II/base/thread_management.h>
 #include <deal.II/base/utilities.h>
 
 #include <deal.II/fe/fe_dgq.h>
